requirements.txt
asgiref==3.2.3
coverage==4.5.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0
django-ckeditor==5.8.0
django-crispy-forms==1.8.1
django-import-export==2.0
django-js-asset==1.2.2
django-material-admin==1.6.7
django-python3-ldap==0.11.3
django-request-vars==1.0.1
django-simple-log==0.3.8
djangorestframework==3.10.3
et-xmlfile==1.0.1
jdcal==1.4.1
ldap3==2.6.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
python-decouple==3.3
pytz==2019.3
PyYAML==5.2
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
